refactor(types): Split mcp/types.py into domain-specific modules
Split the monolithic 1833-line types.py file (53 dataclasses) into 12
focused modules organized by domain:

- board.py: Board analysis types (BoardDimensions, LayerInfo, etc.)
- gerber.py: Gerber export types (GerberFile, GerberExportResult)
- drc.py: DRC violation types (DRCViolation, DRCResult)
- drc_delta.py: DRC delta types (DRCViolationDetail, DRCDeltaInfo)
- session.py: Session management types (SessionInfo, ApplyMoveResult)
- assembly.py: Assembly export types (BOMExportResult, PnPExportResult)
- placement.py: Placement analysis types (PlacementScores, PlacementIssue)
- clearance.py: Clearance measurement types (ClearanceMeasurement)
- routing.py: Routing types (NetRoutingStatus, RouteNetResult)
- intent.py: Intent declaration types (ConstraintInfo, IntentStatus)
- warnings.py: Predictive warning types (PredictiveWarningInfo)

The __init__.py re-exports all types for full backwards compatibility -
existing imports from kicad_tools.mcp.types continue to work unchanged.

Also fixes a pre-existing import bug in design.py where SessionNotFoundError
was imported from the wrong module.

Closes #852
